# openapi.json
{
  "components": {
    "schemas": {
      "CategoryCollection": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            },
            "title": "Results",
            "type": "array"
          },
          "total_categories": {
            "title": "Total Categories",
            "type": "integer"
          }
        },
        "required": [
          "results",
          "total_categories"
        ],
        "title": "CategoryCollection",
        "type": "object"
      },
      "CategorySummary": {
        "properties": {
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "currency": {
            "const": "EUR",
            "default": "EUR",
            "title": "Currency",
            "type": "string"
          },
          "in_stock_count": {
            "title": "In Stock Count",
            "type": "integer"
          },
          "max_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Price"
          },
          "min_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Price"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_count": {
            "title": "Product Count",
            "type": "integer"
          }
        },
        "required": [
          "category_id",
          "name",
          "product_count",
          "in_stock_count",
          "min_price",
          "max_price"
        ],
        "title": "CategorySummary",
        "type": "object"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "recovery_hint": {
            "title": "Recovery Hint",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "recovery_hint"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ProductCollection": {
        "properties": {
          "filters_applied": {
            "additionalProperties": true,
            "title": "Filters Applied",
            "type": "object"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ProductSummary"
            },
            "title": "Results",
            "type": "array"
          },
          "returned_count": {
            "title": "Returned Count",
            "type": "integer"
          },
          "total_matches": {
            "title": "Total Matches",
            "type": "integer"
          }
        },
        "required": [
          "results",
          "total_matches",
          "returned_count",
          "offset",
          "next_offset",
          "filters_applied"
        ],
        "title": "ProductCollection",
        "type": "object"
      },
      "ProductDetailResponse": {
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "recommendation_eligibility": {
            "$ref": "#/components/schemas/RecommendationEligibility"
          }
        },
        "required": [
          "product",
          "recommendation_eligibility"
        ],
        "title": "ProductDetailResponse",
        "type": "object"
      },
      "ProductSummary": {
        "properties": {
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "category_name": {
            "title": "Category Name",
            "type": "string"
          },
          "currency": {
            "const": "EUR",
            "default": "EUR",
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "occasions": {
            "items": {
              "type": "string"
            },
            "title": "Occasions",
            "type": "array"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "recipient": {
            "title": "Recipient",
            "type": "string"
          },
          "shipping_days": {
            "title": "Shipping Days",
            "type": "integer"
          },
          "stock_status": {
            "$ref": "#/components/schemas/StockStatus"
          },
          "subcategory": {
            "title": "Subcategory",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "name",
          "category_id",
          "category_name",
          "subcategory",
          "price",
          "stock_status",
          "shipping_days",
          "recipient",
          "occasions",
          "rating",
          "description"
        ],
        "title": "ProductSummary",
        "type": "object"
      },
      "RecommendationEligibility": {
        "properties": {
          "can_recommend_as_available": {
            "title": "Can Recommend As Available",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "can_recommend_as_available",
          "reason"
        ],
        "title": "RecommendationEligibility",
        "type": "object"
      },
      "StockStatus": {
        "enum": [
          "in_stock",
          "out_of_stock"
        ],
        "title": "StockStatus",
        "type": "string"
      }
    },
    "securitySchemes": {
      "CatalogApiKey": {
        "description": "Secret server-to-server key configured in Indigo. Never place it in a URL.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Deterministic, authenticated catalog tools for the Indigo Product Discovery Agent. Product facts come only from the configured CSV.",
    "title": "Gift Shop Catalog API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/categories": {
      "get": {
        "description": "Use when the user asks what types of gifts can be browsed or has not selected a category. Do not call automatically when a valid category or contextual product request is already known. Returns every compact category summary.",
        "operationId": "get_categories",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid API key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown category or product."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid argument."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Catalog unavailable."
          }
        },
        "security": [
          {
            "CatalogApiKey": []
          }
        ],
        "summary": "List the available gift categories",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/v1/categories/{category_id}/products": {
      "get": {
        "description": "Use when a category_id returned by get_categories is known and the user wants to browse it with hard price, stock, or delivery constraints. Do not use for cross-category natural-language discovery or fetch details for every result.",
        "operationId": "get_products_by_category",
        "parameters": [
          {
            "description": "Exact category_id from get_categories.",
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "description": "Exact category_id from get_categories.",
              "title": "Category Id",
              "type": "string"
            }
          },
          {
            "description": "Optional minimum price in EUR.",
            "in": "query",
            "name": "min_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional minimum price in EUR.",
              "title": "Min Price"
            }
          },
          {
            "description": "Hard maximum price in EUR.",
            "in": "query",
            "name": "max_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hard maximum price in EUR.",
              "title": "Max Price"
            }
          },
          {
            "description": "Hard maximum catalog shipping estimate in days.",
            "in": "query",
            "name": "max_shipping_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 7,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hard maximum catalog shipping estimate in days.",
              "title": "Max Shipping Days"
            }
          },
          {
            "description": "Exclude zero-stock products when true.",
            "in": "query",
            "name": "in_stock_only",
            "required": false,
            "schema": {
              "default": true,
              "description": "Exclude zero-stock products when true.",
              "title": "In Stock Only",
              "type": "boolean"
            }
          },
          {
            "description": "Page size, maximum 10.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "description": "Page size, maximum 10.",
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Zero-based result offset.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Zero-based result offset.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid API key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown category or product."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid argument."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Catalog unavailable."
          }
        },
        "security": [
          {
            "CatalogApiKey": []
          }
        ],
        "summary": "Browse products in one known category",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/v1/products/{product_id}": {
      "get": {
        "description": "Use only when product_id is already known from a collection operation and full facts must be verified. For a product name, call search_products first. Do not use this operation for discovery or catalog enumeration.",
        "operationId": "get_product_details",
        "parameters": [
          {
            "description": "Exact source product_id.",
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "description": "Exact source product_id.",
              "minLength": 1,
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid API key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown category or product."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid argument."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Catalog unavailable."
          }
        },
        "security": [
          {
            "CatalogApiKey": []
          }
        ],
        "summary": "Get full facts for one known product ID",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/v1/search/products": {
      "get": {
        "description": "Use for natural-language or cross-category discovery involving interests, recipient, occasion, intended use, budget, or delivery constraints. Send concise catalog concepts rather than the whole conversation. Use this with in_stock_only=false to resolve an exact product name before requesting detail. Do not use when only the category list or a known product_id is needed.",
        "operationId": "search_products",
        "parameters": [
          {
            "description": "Concise product need, interest, occasion, use, or exact name.",
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "description": "Concise product need, interest, occasion, use, or exact name.",
              "examples": [
                "practical kitchen gift"
              ],
              "maxLength": 200,
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "description": "Optional exact ID from get_categories.",
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional exact ID from get_categories.",
              "title": "Category Id"
            }
          },
          {
            "description": "Recipient such as her or him.",
            "in": "query",
            "name": "recipient",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Recipient such as her or him.",
              "title": "Recipient"
            }
          },
          {
            "description": "Occasion such as housewarming.",
            "in": "query",
            "name": "occasion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Occasion such as housewarming.",
              "title": "Occasion"
            }
          },
          {
            "description": "Minimum EUR price.",
            "in": "query",
            "name": "min_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum EUR price.",
              "title": "Min Price"
            }
          },
          {
            "description": "Hard maximum EUR price.",
            "in": "query",
            "name": "max_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hard maximum EUR price.",
              "title": "Max Price"
            }
          },
          {
            "description": "Hard maximum shipping estimate.",
            "in": "query",
            "name": "max_shipping_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 7,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hard maximum shipping estimate.",
              "title": "Max Shipping Days"
            }
          },
          {
            "description": "Exclude zero-stock products when true.",
            "in": "query",
            "name": "in_stock_only",
            "required": false,
            "schema": {
              "default": true,
              "description": "Exclude zero-stock products when true.",
              "title": "In Stock Only",
              "type": "boolean"
            }
          },
          {
            "description": "Page size, maximum 10.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "description": "Page size, maximum 10.",
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Zero-based result offset.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Zero-based result offset.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing API key."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid API key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown category or product."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid argument."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Catalog unavailable."
          }
        },
        "security": [
          {
            "CatalogApiKey": []
          }
        ],
        "summary": "Search products across the catalog",
        "tags": [
          "Catalog"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "schema-export",
      "url": "https://catalog-api.example.com"
    }
  ]
}